[proofs] [cnf] Handle proofs for optimized clauses in incremental
Since the SAT solver may optimize the assertion level of propagation
explanations and of added clauses, we need to be able to maintain proofs across
context-popping. Not doing this can lead to open proofs. To do this the proof
cnf stream will use an OptClausesManager to re-add proofs of facts inserted at
optimized levels.

This commit adds the notification methods that the SAT solver can call to notify
when a clause was added at an optimized level.