[proofs] [sat] Have SAT solver communicate proof utils of optimized clauses
This way both the SAT proof manager and the Proof Cnf Stream can properly track
proofs for clauses added at assertion levels below their original user level.

This commit, besides some minor tweakes, adds the hooks in the places where
a clause can be created with an optimized level and we need to track:
- when the explanation of a propagation is requested
- when a clause/lemma is added
- when a lemma is added (from the backlog of lemmas added when Minisat was busy)

For the resolution, since the clause is passed to the SAT proof manager, it can
by itself determine whether the result of the resolution chain is having its
level optimized.

When a conflict is derived