Send update/nth on unit sequences to array solver
This commit changes our policy of which terms we send to the array core
solver. Previously, we were not sending `update` and `nth` terms that
were applied to unit sequences. This was incorrect, because for example
the split `STRINGS_ARRAY_EQ_SPLIT` has to be applied to `nth` of units,
too, to guarantee model soundness: For example, if we have a disequality
`nth(t1, i) != nth(unit(v), i)`, then we have to reason about the
(dis)equality of `t1` and `unit(v)`, because of the UF for the
out-of-bounds case of `nth` (i.e., we have to reason that `Uf(t1, i) !=
Uf(unit(v), i)`, which requires `t1 != unit(v)`). Splitting on `t1 =
unit(v)` is done by `STRINGS_ARRAY_EQ_SPLIT`.