[Strings] Register terms more lazily in EE
Until now, we have been registering terms in the equality engine of the
string solver eagerly during pre-registration. At the same time, the
shared solver is only considering shared terms in facts that have been
asserted. As a result, it could happen that the string solver was asking
for a value of a shared term during model construction that the shared
solver did not get notified of, leading to an assertion failure. This
commit changes when terms are registered with the equality engine.