[CI] Make building static/shared configurable
This commit changes our CI workflow to actually pass through
`build-shared` and `build-static` to the `configure-and-build` action
and fixes how the condition is checked there. If we just pass through
the inputs, then the condition `"${{ inputs.build-shared }}" != "true"`
would be `true` (i.e., the step would be skipped) if
`build-shared`/`build-static` are not set explicitly in `ci.yml` because
of the default values for `build-shared`/`build-static` in `ci.yml`.