[BV] Fix strategy for rewriting `bvnot`
Fixes #8240. The currenty strategy was returning `REWRITE_DONE`
incorrectly, e.g., when we had a term like `(bvnot (bvnot (bvnot #b0)))`
because `EvalNot` was applied before `NotIdemp`, so we did not evaluate
the innermost `bvnot` after removing the double `bvnot`. The commit
fixes the strategy s.t. we are allowed to return `REWRITE_DONE`.