Ignore `CMAKE_SYSROOT` when cross-compiling
This commit changes our build system to ignore `CMAKE_SYSROOT` for
`find_file()`, `find_path()`, and `find_library()` when cross-compiling.
Previously, it could happen that, e.g., when we were cross-compiling a
static build for Windows, CMake would pick up `/usr/lib/libgmp.a`
instead of cross-compiling GMP. The comments in the `Toolchain-*.cmake`
files suggest that the new behavior was the intended behavior all along.