[API] Remove redundant version of `mkFunctionSort`
`mkFunctionSort` that takes two sorts as arguments is redundant, because
the first argument is equivalent to a vector of size one passed to the
other overload of `mkFunctionSort`. This commit removes the method from
the C++ API but keeps the existing semantics for the Java and Python
bindings for convenience and consistency with, e.g. `mkTerm`.